修复L2 Codex

+++ frameworks/resmgr/src/raw_file_manager.cpp
@@ -177,8 +177,8 @@ const char *OH_ResourceManager_GetRawFileName(RawDir *rawDir, int index)
     if (rawDir == nullptr || index < 0) {
         return nullptr;
     }
-    int rawFileCount = rawDir->fileNameCache.names.size();
-    if (rawFileCount == 0 || index >= rawFileCount) {
+    uint32_t rawFileCount = rawDir->fileNameCache.names.size();
+    if (rawFileCount == 0 || index >= static_cast<int>(rawFileCount)) {
         return nullptr;
     }
     return rawDir->fileNameCache.names[index].c_str();
